+++ R/covr.R
@@ -1,8 +1,8 @@
 #' covr: Test coverage for packages
 #'
 #' covr tracks and reports code coverage for your package and (optionally)
-#' upload the results to a coverage service like 'Codecov' <http://codecov.io> or
-#' 'Coveralls' <http://coveralls.io>. Code coverage is a measure of the amount of
+#' upload the results to a coverage service like 'Codecov' <https://codecov.io> or
+#' 'Coveralls' <https://coveralls.io>. Code coverage is a measure of the amount of
 #' code being exercised by a set of tests. It is an indirect measure of test
 #' quality and completeness. This package is compatible with any testing
 #' methodology or framework and tracks coverage of both R code and compiled
@@ -15,14 +15,72 @@
 #' # If run with no arguments `report()` implicitly calls `package_coverage()`
 #' report()
 #' ```
+#'
+#' @section Package options:
+#'
+#' `covr` uses the following [options()] to configure behaviour:
+#'
+#' \itemize{
+#'   \item `covr.covrignore`: A filename to use as an ignore file,
+#'     listing glob-style wildcarded paths of files to ignore for coverage
+#'     calculations. Defaults to the value of environment variable
+#'     `COVR_COVRIGNORE`, or `".covrignore"`  if the neither the option nor the
+#'     environment variable are set.
+#'
+#'   \item `covr.exclude_end`: Used along with `covr.exclude_start`, an optional
+#'     regular expression which ends a line-exclusion region. For more
+#'     details, see `?exclusions`.
+#'
+#'   \item `covr.exclude_pattern`: An optional line-exclusion pattern. Lines
+#'     which match the pattern will be excluded from coverage. For more details,
+#'     see `?exclusions`.
+#'
+#'   \item `covr.exclude_start`: Used along with `covr.exclude_end`, an optional
+#'     regular expression which starts a line-exclusion region. For more
+#'     details, see `?exclusions`.
+#'
+#'   \item `covr.filter_non_package`: If `TRUE` (the default behavior), coverage
+#'     of files outside the target package are filtered from coverage output.
+#'
+#'   \item `covr.fix_parallel_mcexit`:
+#'
+#'   \item `covr.flags`:
+#'
+#'   \item `covr.gcov`: If the appropriate gcov version is not on your path you
+#'     can use this option to set the appropriate location. If set to "" it will
+#'     turn off coverage of compiled code.
+#'
+#'   \item `covr.gcov_additional_paths`:
+#'
+#'   \item `covr.gcov_args`:
+#'
+#'   \item `covr.icov`:
+#'
+#'   \item `covr.icov_args`:
+#'
+#'   \item `covr.icov_flags`:
+#'
+#'   \item `covr.icov_prof`:
+#'
+#'   \item `covr.rstudio_source_markers`: A logical value. If `TRUE` (the
+#'     default behavior), source markers are displayed within the RStudio IDE
+#'     when using `zero_coverage`.
+#'
+#'   \item `covr.record_tests`: If `TRUE` (default `NULL`), record a listing of
+#'     top level test expressions and associate tests with `covr` traces
+#'     evaluated during the test's execution. For more details, see
+#'     `?covr.record_tests`.
+#'
+#'   \item `covr.showCfunctions`:
+#' }
+#'
+#'
 "_PACKAGE"
 
 #' @import methods
 #' @importFrom stats aggregate na.omit na.pass setNames
 #' @importFrom utils capture.output getSrcFilename relist str head
 NULL
-
-rex::register_shortcuts("covr")
 
 the <- new.env(parent = emptyenv())
 
@@ -49,12 +107,36 @@ save_trace <- function(directory) {
   saveRDS(.counters, file = tmp_file)
 }
 
+#' Convert a counters object to a coverage object
+#' 
+#' @param counters An environment of covr trace results to convert to a coverage
+#'   object. If `counters` is not provided, the `covr` namespace value
+#'   `.counters` is used. 
+#' @param ... Additional attributes to include with the coverage object.
+#'
+as_coverage <- function(counters = NULL, ...) {
+  if (missing(counters))
+    counters <- .counters
+
+  counters <- as.list(counters)
+
+  # extract optional tests 
+  tests <- counters$tests
+  counters$tests <- NULL
+
+  structure(counters, tests = tests, ..., class = "coverage")
+}
+
 #' Calculate test coverage for a specific function.
 #'
 #' @param fun name of the function.
 #' @param code expressions to run.
 #' @param env environment the function is defined in.
 #' @param enc the enclosing environment which to run the expressions.
+#' @examples
+#' add <- function(x, y) { x + y }
+#' function_coverage(fun = add, code = NULL) # 0% coverage
+#' function_coverage(fun = add, code = add(1, 2) == 3) # 100% coverage
 #' @export
 function_coverage <- function(fun, code = NULL, env = NULL, enc = parent.frame()) {
   if (is.function(fun)) {
@@ -78,8 +160,12 @@ function_coverage <- function(fun, code = NULL, env = NULL, enc = parent.frame()
   })
 
   replace(replacement)
-  eval(code, enc)
-  structure(as.list(.counters), class = "coverage")
+
+  withr::with_envvar(c("R_COVR" = "true"),
+    eval(code, enc)
+  )
+
+  as_coverage(as.list(.counters))
 }
 
 #' Calculate test coverage for sets of files
@@ -93,6 +179,12 @@ function_coverage <- function(fun, code = NULL, env = NULL, enc = parent.frame()
 #'   functions
 #' @param parent_env The parent environment to use when sourcing the files.
 #' @inheritParams package_coverage
+#' @examples
+#' # For the purpose of this example, save code containing code and tests to files
+#' cat("add <- function(x, y) { x + y }", file="add.R")
+#' cat("add(1, 2) == 3", file="add_test.R")
+#' # Use file_coverage() to calculate test coverage
+#' file_coverage(source_files = "add.R", test_files = "add_test.R")
 #' @export
 file_coverage <- function(
   source_files,
@@ -114,10 +206,12 @@ file_coverage <- function(
     clear_counters()
   })
 
-  lapply(test_files,
-    sys.source, keep.source = TRUE, envir = env)
+  withr::with_envvar(c("R_COVR" = "true"),
+    lapply(test_files,
+      sys.source, keep.source = TRUE, envir = env)
+  )
 
-  coverage <- structure(as.list(.counters), class = "coverage")
+  coverage <- as_coverage(.counters)
 
   exclude(coverage,
     line_exclusions = line_exclusions,
@@ -134,6 +228,10 @@ file_coverage <- function(
 #' @param test_code A character vector of test code
 #' @inheritParams file_coverage
 #' @param ... Additional arguments passed to [file_coverage()]
+#' @examples
+#' source <- "add <- function(x, y) { x + y }"
+#' test <- "add(1, 2) == 3"
+#' code_coverage(source, test)
 #' @export
 code_coverage <- function(
    source_code,
@@ -169,10 +267,12 @@ environment_coverage <- function(
     clear_counters()
   })
 
-  lapply(test_files,
-    sys.source, keep.source = TRUE, envir = exec_env)
+  withr::with_envvar(c("R_COVR" = "true"),
+    lapply(test_files,
+      sys.source, keep.source = TRUE, envir = exec_env)
+  )
 
-  coverage <- structure(as.list(.counters), class = "coverage")
+  coverage <- as_coverage(.counters)
 
   exclude(coverage,
     line_exclusions = line_exclusions,
@@ -194,7 +294,7 @@ environment_coverage <- function(
 #'
 #' #ifdef unix
 #' Parallelized code using \pkg{parallel}'s [mcparallel()] needs to
-#' be use a patched `parallel:::mcexit`. This is done automatically if the
+#' use a patched `parallel:::mcexit`. This is done automatically if the
 #' package depends on \pkg{parallel}, but can also be explicitly set using the
 #' environment variable `COVR_FIX_PARALLEL_MCEXIT` or the global option
 #' `covr.fix_parallel_mcexit`.
@@ -221,6 +321,10 @@ environment_coverage <- function(
 #' @param ... Additional arguments passed to [tools::testInstalledPackage()].
 #' @param exclusions \sQuote{Deprecated}, please use \sQuote{line_exclusions} instead.
 #' @param pre_clean whether to delete all objects present in the src directory before recompiling
+#' @param install_path The path the instrumented package will be installed to
+#'   and tests run in. By default it is a path in the R sessions temporary
+#'   directory. It can sometimes be useful to set this (along with `clean =
+#'   FALSE`) to help debug test failures.
 #' @seealso [exclusions()] For details on excluding parts of the
 #' package from the coverage calculations.
 #' @export
@@ -233,6 +337,7 @@ package_coverage <- function(path = ".",
                              line_exclusions = NULL,
                              function_exclusions = NULL,
                              code = character(),
+                             install_path = temp_file("R_LIBS"),
                              ...,
                              exclusions, pre_clean=TRUE) {
 
@@ -269,8 +374,7 @@ package_coverage <- function(path = ".",
     return(res)
   }
 
-  tmp_lib <- temp_file("R_LIBS")
-  dir.create(tmp_lib)
+  dir.create(install_path)
 
   flags <- getOption("covr.flags")
 
@@ -294,7 +398,7 @@ package_coverage <- function(path = ".",
       clean_objects(pkg$path)
       clean_gcov(pkg$path)
       clean_parse_data()
-      unlink(tmp_lib, recursive = TRUE)
+      unlink(install_path, recursive = TRUE)
     }, add = TRUE)
   }
 
@@ -304,7 +408,7 @@ package_coverage <- function(path = ".",
   # install the package in a temporary directory
   withr::with_makevars(flags, assignment = "+=",
     utils::install.packages(repos = NULL,
-                            lib = tmp_lib,
+                            lib = install_path,
                             pkg$path,
                             type = "source",
                             INSTALL_opts = c("--example",
@@ -316,39 +420,40 @@ package_coverage <- function(path = ".",
                             quiet = quiet))
 
   # add hooks to the package startup
-  add_hooks(pkg$package, tmp_lib,
+  add_hooks(pkg$package, install_path,
     fix_mcexit = should_enable_parallel_mcexit_fix(pkg))
 
-  libs <- env_path(tmp_lib, .libPaths())
+  libs <- env_path(install_path, .libPaths())
 
   withr::with_envvar(
     c(R_DEFAULT_PACKAGES = "datasets,utils,grDevices,graphics,stats,methods",
       R_LIBS = libs,
       R_LIBS_USER = libs,
       R_LIBS_SITE = libs,
-      R_COVR = "true"), {
+      R_COVR = "true",
+      R_TESTS = file.path(R.home("share"), "R", "tests-startup.R")), {
 
 
     withCallingHandlers({
       if ("vignettes" %in% type) {
         type <- type[type != "vignettes"]
-        run_vignettes(pkg, tmp_lib)
+        run_vignettes(pkg, install_path)
       }
 
-      out_dir <- file.path(tmp_lib, pkg$package)
+      out_dir <- file.path(install_path, pkg$package)
       if ("examples" %in% type) {
         type <- type[type != "examples"]
         # testInstalledPackage explicitly sets R_LIBS="" on windows, and does
         # not restore it after, so we need to reset it ourselves.
         withr::with_envvar(c(R_LIBS = Sys.getenv("R_LIBS")), {
-          result <- tools::testInstalledPackage(pkg$package, outDir = out_dir, types = "examples", lib.loc = tmp_lib, ...)
+          result <- tools::testInstalledPackage(pkg$package, outDir = out_dir, types = "examples", lib.loc = install_path, ...)
           if (result != 0L) {
             show_failures(out_dir)
           }
         })
       }
       if ("tests" %in% type) {
-        result <- tools::testInstalledPackage(pkg$package, outDir = out_dir, types = "tests", lib.loc = tmp_lib, ...)
+        result <- tools::testInstalledPackage(pkg$package, outDir = out_dir, types = "tests", lib.loc = install_path, ...)
         if (result != 0L) {
           show_failures(out_dir)
         }
@@ -356,34 +461,44 @@ package_coverage <- function(path = ".",
 
       # We always run the commands file (even if empty) to load the package and
       # initialize all the counters to 0.
-      run_commands(pkg, tmp_lib, code)
+      run_commands(pkg, install_path, code)
     },
     message = function(e) if (quiet) invokeRestart("muffleMessage") else e,
     warning = function(e) if (quiet) invokeRestart("muffleWarning") else e)
     })
 
   # read tracing files
-  trace_files <- list.files(path = tmp_lib, pattern = "^covr_trace_[^/]+$", full.names = TRUE)
+  trace_files <- list.files(path = install_path, pattern = "^covr_trace_[^/]+$", full.names = TRUE)
   coverage <- merge_coverage(trace_files)
   if (!uses_icc()) {
-    res <- run_gcov(pkg$path, quiet = quiet)
+    res <- run_gcov(pkg$path, quiet = quiet, clean = clean)
   } else {
     res <- run_icov(pkg$path, quiet = quiet)
   }
 
-  coverage <- structure(c(coverage, res),
-      class = "coverage",
-      package = pkg,
-      relative = relative_path)
+  coverage <- as_coverage(
+    c(coverage, res),
+    package = pkg,
+    relative = relative_path
+  )
 
   if (!clean) {
-    attr(coverage, "library") <- tmp_lib
+    attr(coverage, "library") <- install_path
   }
 
-  coverage <- filter_non_package_files(coverage)
+  if (getOption("covr.filter_non_package", TRUE)) {
+    coverage <- filter_non_package_files(coverage)
+  }
 
-  # Exclude both RcppExports to avoid redundant coverage information
-  line_exclusions <- c("src/RcppExports.cpp", "R/RcppExports.R", line_exclusions, parse_covr_ignore())
+  # Exclude generated files from Rcpp and cpp11 to avoid redundant coverage information
+  line_exclusions <- c(
+    "src/RcppExports.cpp",
+    "R/RcppExports.R",
+    "src/cpp11.cpp",
+    "R/cpp11.R",
+    line_exclusions,
+    parse_covr_ignore()
+  )
 
   exclude(coverage,
     line_exclusions = line_exclusions,
@@ -434,29 +549,46 @@ show_failures <- function(dir) {
 # merge multiple coverage files together. Assumes the order of coverage lines
 # is the same in each object, this should always be the case if the objects are
 # from the same initial library.
-merge_coverage <- function(files) {
-  nfiles <- length(files)
-  if (nfiles == 0) {
+merge_coverage <- function(x) {
+  UseMethod("merge_coverage")
+}
+
+merge_coverage.character <- function(files) {
+  coverage_objs <- lapply(files, function(f) {
+    as.list(suppressWarnings(readRDS(f)))
+  })
+  merge_coverage(coverage_objs)
+}
+
+merge_coverage.list <- function(coverage_objs) {
+  if (length(coverage_objs) == 0) {
     return()
   }
 
-  x <- suppressWarnings(readRDS(files[1]))
-  x <- as.list(x)
-  if (nfiles == 1) {
-    return(x)
-  }
-
+  x <- coverage_objs[[1]]
   names <- names(x)
-  for (i in 2:nfiles) {
-    y <- suppressWarnings(readRDS(files[i]))
+
+  for (y in tail(coverage_objs, -1L)) {
+    # align tests from coverage objects
+    test_idx <- match(names(y$tests), Filter(nchar, names(x$tests)))
+    new_test_idx <- if (!length(test_idx)) seq_along(y$tests) else which(is.na(test_idx))
+    test_idx[new_test_idx] <- length(x$tests) + seq_along(new_test_idx)
+
+    # append any tests that we haven't encountered in previous objects
+    x$tests <- append(x$tests, y$tests[new_test_idx])
+    y$tests <- NULL
+
     for (name in intersect(names, names(y))) {
       x[[name]]$value <- x[[name]]$value + y[[name]]$value
+      y[[name]]$tests[,1] <- test_idx[y[[name]]$tests[,1]]
+      x[[name]]$tests <- rbind(x[[name]]$tests, y[[name]]$tests)
     }
+
     for (name in setdiff(names(y), names)) {
       x[[name]] <- y[[name]]
     }
+
     names <- union(names, names(y))
-    y <- NULL
   }
 
   x
@@ -530,13 +662,16 @@ run_commands <- function(pkg, lib, commands) {
 # @param pkg_name name of the package to add hooks to
 # @param lib the library path to look in
 # @param fix_mcexit whether to add the fix for mcparallel:::mcexit
-add_hooks <- function(pkg_name, lib, fix_mcexit = FALSE) {
+add_hooks <- function(pkg_name, lib, fix_mcexit = FALSE,
+  record_tests = isTRUE(getOption("covr.record_tests", FALSE))) {
+
   trace_dir <- paste0("Sys.getenv(\"COVERAGE_DIR\", \"", lib, "\")")
 
   load_script <- file.path(lib, pkg_name, "R", pkg_name)
   lines <- readLines(file.path(lib, pkg_name, "R", pkg_name))
   lines <- append(lines,
-    c("setHook(packageEvent(pkg, \"onLoad\"), function(...) covr:::trace_environment(ns))",
+    c(paste0("setHook(packageEvent(pkg, \"onLoad\"), function(...) options(covr.record_tests = ", record_tests, "))"),
+      "setHook(packageEvent(pkg, \"onLoad\"), function(...) covr:::trace_environment(ns))",
       paste0("reg.finalizer(ns, function(...) { covr:::save_trace(", trace_dir, ") }, onexit = TRUE)")),
     length(lines) - 1L)
 
@@ -550,4 +685,18 @@ add_hooks <- function(pkg_name, lib, fix_mcexit = FALSE) {
 #' @export
 `[.coverage` <- function(x, ...) {
   structure(NextMethod(), class = "coverage")
+}
+
+#' Determine if code is being run in covr
+#'
+#' covr functions set the environment variable `R_COVR` when they are running.
+#' [in_covr()] returns `TRUE` if this environment variable is set and `FALSE`
+#' otherwise.
+#' @export
+#' @examples
+#' if (require(testthat)) {
+#'   testthat::skip_if(in_covr())
+#' }
+in_covr <- function() {
+  identical(Sys.getenv("R_COVR"), "true")
 }